requirements.txt
aiogram==3.*
aiohttp
python-dotenv
requests
googletrans
matplotlib
